random ws request ID long type to int

+++ library/src/main/java/org/whispersystems/websocket/WebSocketClient.java
@@ -99,7 +99,7 @@ public class WebSocketClient {
   }
 
   private long generateRequestId() {
-    return Math.abs(new SecureRandom().nextLong());
+    return Math.abs(new SecureRandom().nextInt());
   }
 
 }
